% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abracadabra.R
\name{abracadabraAlacazam_plotmyELISA_withthisProgram}
\alias{abracadabraAlacazam_plotmyELISA_withthisProgram}
\title{Basic ELISA plotting}
\usage{
abracadabraAlacazam_plotmyELISA_withthisProgram(
  y,
  title,
  color,
  facet_by,
  palette
)
}
\arguments{
\item{y}{Data you'd like to put on the y-axis. You can enter "aldo", "viability", or "aldo_norm". Must be in quotation marks.}

\item{title}{Title for your plot. Must be in quotation marks.}

\item{color}{Specifies the variable by which you'd like to subset boxes (ie. OD$dox, OD$cell, etc.). Do not use quotation marks.}

\item{facet_by}{Specifies the variable by which you'd like to facet (ie. OD$cell, etc.). Do not use quotation marks.}

\item{palette}{Specifies the RColorBrewer palette you'd like to use. Recommend "Dark2" or "Set1". Defaults to "Set1". Use quotation marks.}
}
\description{
This function generates a plot of your ELISA data with basic customization features. The plot separates data by 2-3 variables. For example, you could plot different colored bars based on +/-dox and facet the plot by cell type.
}
\details{
This function is intended to be a quick and basic plotting tool. If you need more advanced data visualization, it might not serve all the purposes you need.

Important: the x axis always displays +/- stim. y, title, and palette must be entered in quotation marks and must exactly match the options available. color and facet_by must be in dataframe$columnname format and must not be in quotation marks.
}
\examples{
abracadabraAlacazam_plotmyELISA_withthisProgram(y = "aldo", title = "example title", color = OD$dox, facet_by = OD$cell, palette = "Dark2")
}
